refactor: 변경된 타입에 맞게 수정, 스타일 변경
