+++ src/apis/profile/queries.ts
@@ -17,8 +17,6 @@ export const useGetProfile = () =>
 
       return { ...data, introduce: defaultIntroduce, birth: formattedBirth };
     },
-    staleTime: Infinity,
-    cacheTime: Infinity,
   });
 
 export const useGetProfileById = (userId: number) =>
@@ -36,8 +34,6 @@ export const useGetProfileById = (userId: number) =>
 
       return { ...data, introduce: defaultIntroduce, birth: formattedBirth };
     },
-    staleTime: Infinity,
-    cacheTime: Infinity,
   });
 
 export const useGetPraises = () => useSuspenseQuery(Keys.getPraises(), getPraises);
