+++ src/app/(main)/profile/setting/components/ProfileSection.client.tsx
@@ -9,18 +9,22 @@ import Link from 'next/link';
 
 export default function ProfileSection() {
   const { data: profileData } = useGetProfile();
-  const { imageUrl, name, reliability } = profileData;
+  const { imageUrl, nickname, reliabilityLevel } = profileData;
 
   return (
     <section className="px-20 py-8">
       <Flex align="center" className="gap-12 rounded-8 bg-sub px-16 py-12">
         <Avatar imageUrl={imageUrl} />
         <div>
-          <p className="text-paragraph-1">{name}</p>
+          <p className="text-paragraph-1">{nickname}</p>
           <Spacing direction="horizontal" size={2} />
           <p className="flex text-caption text-sign-tertiary">
-            <Icon id={`16-reliability-${reliability.toLocaleLowerCase()}`} width={16} height={16} />
-            {reliability}
+            <Icon
+              id={`16-reliability-${reliabilityLevel.toLocaleLowerCase()}`}
+              width={16}
+              height={16}
+            />
+            {reliabilityLevel}
           </p>
         </div>
         <Link href="/profile/setting/edit" className="ml-auto">
